RouteRuleToken model: Integer -> BigInteger (fix INSERT overflow)
User restart vẫn thấy WARNING migrations - route_rule_tokens skipped:
NumericValueOutOfRange. Root cause: model dùng Integer, SQLAlchemy generate
INSERT (rule_id::INTEGER) cast → overflow vì rule_id values từ CRDB
unique_rowid > 2^31.

Fix: change model RouteRuleToken.id + rule_id sang BigInteger. SQLAlchemy
sẽ generate ::BIGINT cast correct.

User đã ALTER DB column sang BIGINT thủ công. Sau pull + restart, model
match DB → backfill thành công.

+++ backend/models.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime, date
 
-from sqlalchemy import Boolean, Date, DateTime, Float, Integer, String, Text, ForeignKey, UniqueConstraint
+from sqlalchemy import BigInteger, Boolean, Date, DateTime, Float, Integer, String, Text, ForeignKey, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from database import Base
@@ -416,8 +416,11 @@ class RouteRuleToken(Base):
     __tablename__ = "route_rule_tokens"
     __table_args__ = (UniqueConstraint("rule_id", "token", name="uq_rule_token"),)
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    rule_id: Mapped[int] = mapped_column(Integer, ForeignKey("route_keyword_rules.id", ondelete="CASCADE"), index=True)
+    # BigInteger vì rule_id reference route_keyword_rules.id có thể > 2^31
+    # (CRDB unique_rowid sinh ~1.18e18). SQLAlchemy Integer cast INSERT thành
+    # INTEGER → NumericValueOutOfRange. BigInteger map sang BIGINT correct.
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    rule_id: Mapped[int] = mapped_column(BigInteger, ForeignKey("route_keyword_rules.id", ondelete="CASCADE"), index=True)
     token: Mapped[str] = mapped_column(String(128), index=True)
